requirements.txt
streamlit
numpy
pandas
joblib
scikit-learn
xgboost
tensorflow-macos
tensorflow-metal
shap
scipy
